requirements.txt
yfinance==0.1.54
lxml==4.5.1
numpy==1.19.0
requests==2.24.0
google-api-python-client==1.9.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
